Finished the proof reading (#1)
* Update 2019-04-15-an-intro-to-dpdk-for-mortals.md

I edited out and replaced some things that felt wrong, while to preserve the intended meaning. Although, this way I might go overboard
(to be continued)

* Update 2019-04-15-an-intro-to-dpdk-for-mortals.md

(to be continued)

* Update 2019-04-15-an-intro-to-dpdk-for-mortals.md

(to be continued)

* Update 2019-04-15-an-intro-to-dpdk-for-mortals.md

(still to be continued)

* Update 2019-04-15-an-intro-to-dpdk-for-mortals.md

And I reached the end

* dpdk upper case

Because I noticed that at some places the DPDK is written in lower case, in this commit I change that to upper case wherever that appears appropriate from the context. (for instance, the lower case is used in file names and the upper case is used in the actual text)